requirements.txt
whoosh==2.7.4
kafka-python==2.0.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
cachetools==5.3.2
fastapi==0.109.1
uvicorn==0.27.1